Refactor and clean
