+++ baremaps-cli/src/main/java/org/apache/baremaps/cli/Baremaps.java
@@ -26,9 +26,9 @@ import java.util.concurrent.Callable;
 import org.apache.baremaps.cli.Baremaps.VersionProvider;
 import org.apache.baremaps.cli.database.Database;
 import org.apache.baremaps.cli.geocoder.Geocoder;
+import org.apache.baremaps.cli.hillshade.Hillshade;
 import org.apache.baremaps.cli.iploc.IpLoc;
 import org.apache.baremaps.cli.map.Map;
-import org.apache.baremaps.cli.raster.Raster;
 import org.apache.baremaps.cli.workflow.Workflow;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.core.config.Configurator;
@@ -49,7 +49,7 @@ import picocli.CommandLine.Option;
         Map.class,
         Geocoder.class,
         IpLoc.class,
-        Raster.class
+        Hillshade.class
     },
     sortOptions = false)
 @SuppressWarnings("squid:S106")
