+++ baremaps-core/pom.xml
@@ -104,6 +104,10 @@ limitations under the License.
     </dependency>
     <dependency>
       <groupId>org.apache.baremaps</groupId>
+      <artifactId>baremaps-raster</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.baremaps</groupId>
       <artifactId>baremaps-testing</artifactId>
     </dependency>
     <dependency>
