+++ baremaps-raster/src/main/java/org/apache/baremaps/raster/ContourTracer.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.baremaps.raster.elevation;
+package org.apache.baremaps.raster;
 
 import java.util.ArrayList;
 import java.util.Arrays;
