+++ baremaps-raster/src/test/java/org/apache/baremaps/raster/MartiniTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.baremaps.raster.martini;
+package org.apache.baremaps.raster;
 
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
